:recycle: refactor(select): extract helpers, hooks, parts and renderOptions into separate modules
